Refactor : ApiClient
js -> ts로 변경
