+++ front/src/api/FeedbackService.ts
@@ -1,4 +1,4 @@
-import ApiClient from "./ApiClient";
+import ApiClient from "./ApiClient.ts";
 
 export interface FeedbackPayload {
   feedbackContent: string;
